attendance-system/requirements.txt
opencv-python-headless>=4.7.0
pyzbar>=0.1.9
python-dotenv>=1.0.0
pymongo>=4.5.0
